Adds basic installation routine
This adds a simple installation script, currently only for Bash, that
will do basic setup of the machine it runs on, like adding aliases,
adding helper functions, adding exports and so on. This is just a start,
other things need to be added later, e.g. installing software and so on.